Add training part for image segmentation - Add `create_color_legend` to show legend for class colors
*MODULE CHANGES to `machine_learning.utils`:
- move command-related functions into `command_utils.py`
- move callbacks into `callbacks.py`
